+++ src/domains/review/components/ReviewWriteForm/CommentSection/CompleteButtonSection.tsx
@@ -11,13 +11,11 @@ const CompleteButtonSection = () => {
   } = useFormContext<IReviewWriteForm>();
 
   return (
-    <div className="fixed z-bottomButton left-1/2 -translate-x-1/2 bottom-0 w-full max-w-[600px] bg-white">
-      <PaddingWrapper>
-        <ButtonNewver color="black" className="w-full" type="submit" disabled={!isValid}>
-          완료
-        </ButtonNewver>
-      </PaddingWrapper>
-    </div>
+    <PaddingWrapper>
+      <ButtonNewver color="black" className="w-full" type="submit" disabled={!isValid}>
+        완료
+      </ButtonNewver>
+    </PaddingWrapper>
   );
 };
